all done

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: lcorinna <lcorinna@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/10/11 16:03:47 by lcorinna          #+#    #+#             */
-/*   Updated: 2021/10/25 22:45:33 by lcorinna         ###   ########.fr       */
+/*   Updated: 2021/11/13 18:24:02 by lcorinna         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int	ft_strlen(const char *r)
+size_t	ft_strlen(const char *r)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (r[i])
